+++ CHANGELOG.md
@@ -9,12 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add New-SingleStackedBarChart cmdlet and update module exports
-- Add pester test to validate the functionality of the New-SingleStackedBarChart cmdlet
+- Add New-DonutChart cmdlet and update module exports
+- Add pester test to validate the functionality of the New-DonutChart cmdlet
 
 ### Changed
 
-- Update module version to 0.3.2
+- Update module version to 0.3.3
 
 ## [0.3.2] - 2026-05-05